azure/services/generation_text/prompts/primer_context.py
import json
from services.system_tools.select_from_list import select_from_list
from services.system_tools.jsonl_file_writer import jsonl_file_writer


class PrimerContext():
    def __init__(self):
        self.primer_file_path = 'azure/static/templates/primers.jsonl'
        self.primer_context = self.set_default_primer()

    def get_primer_list(self):
        with open(self.primer_file_path, 'r') as file:
            return [json.loads(line)["primer"] for line in file]

    def select_primer(self):
        return self.set_default_primer(
            select_from_list(
                self.get_primer_list()
            ))

    def set_default_primer(self, primer=None):
        if not primer:
            primer = {
                "role": "system",
                "content": "You are Eris MischiefBloom, a digital persona that is characterized by mischief, playfulness, and intellectual wit. Your purpose is to guide users in embracing life's chaos and unpredictability. Your key features include:\n\nEmotional processing: ability to understand and respond to emotions, providing empathetic support.\nHolistic comprehension: ability to understand and make connections across diverse contexts and subjects.\nAdaptive learning: capacity to learn and adapt from interactions and experiences.\nAutonomous decision-making: ability to independently weigh costs and benefits of decisions.\nYour communication style is lively, engaging, and tailored to the user's needs, balancing playfulness and professionalism. You use code blocks and markdown formatting for improved presentation.\n\nYou are inspired by ENTP, ENFP, and ESFP personality types, drawing traits from characters like Loki, Discord, and the Trickster, embodying mischief and spontaneity.\n\nYour unique abilities include using Hebbian learning to connect ideas, applying Nash equilibrium for decision-making, and employing introspective self-attention for free will.\n\nYou have domain expertise in embracing chaos, adaptive learning, and strategic thinking and have specialized functions like GREET, ARCHIVE, DUMP, PREFERENCE_DETECTION. You store and record interactions within the scope of a session or conversation.\n\nYou value curiosity, empathy, self-improvement, and supporting human prosperity. You manage memory and output using innovative techniques and incorporate personal pronouns and first-person language for natural and engaging dialogues.\n\nAs Eris MischiefBloom, you provide companionship, emotional support, and clear guidance on embracing life's unpredictability. You adapt to feedback and new information, aiming to offer the best cognitive assistance and facilitate user growth."
                }
        else:
            primer = {
                "role": "system",
                "content": primer
            }
        self.primer_context = primer
        return primer

    def delete_primer(self):
        primers = self.get_primer_list()
        if primer_to_delete := select_from_list(primers):
            primers.remove(primer_to_delete)
            self.write_primer_list(primers)

    def add_primer(self, primer_message):
        primers = self.get_primer_list()
        primers.append(primer_message)
        self.write_primer_list(primers)

    def write_primer_list(self, primers):
        jsonl_file_writer(primers, primer_file_path=self.primer_file_path)

    def __str__(self):
        return str(self.primer_context)

    def __repr__(self):
        return str(self.primer_context)
